requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
cloudinary==1.34.0
cryptography==41.0.3
dj-database-url==2.1.0
Django==4.2.4
django-cors-headers==4.2.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==10.0.0
pycparser==2.21
pycryptodome==3.18.0
PyJWT==2.8.0
pytz==2023.3
rest-framework-simplejwt==0.0.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
whitenoise==6.5.0